# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users Api",
    "description": "Users Api",
    "version": "1.0.0"
  },
  "host": "cse341-final-i3fm.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "paths": {
...
    "/teacher/": {
      "get": {
        "tags": ["teachers"],
        "description": "Get all teachers",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109ca"
                  },
                  "name": {
                    "type": "string",
                    "example": "John Smith"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Mathematics"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["teachers"],
        "summary": "Create a new teacher",
        "description": "Creates a new teacher in the database.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Teacher object that needs to be added",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "email", "subject"],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Smith"
                },
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                },
                "subject": {
                  "type": "string",
                  "example": "Mathematics"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Teacher created successfully"
          },
          "422": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/teacher/{id}": {
      "get": {
        "tags": ["teachers"],
        "description": "Get a single teacher by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Teacher ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher found",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "60d0fe4f5311236168a109ca"
                },
                "name": {
                  "type": "string",
                  "example": "John Smith"
                },
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                },
                "subject": {
                  "type": "string",
                  "example": "Mathematics"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Teacher not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["teachers"],
        "summary": "Update an existing teacher",
        "description": "Updates a teacher by ID.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the teacher to update",
            "example": "60d0fe4f5311236168a109ca"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated teacher object",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "email", "subject"],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Jane Doe"
                },
                "email": {
                  "type": "string",
                  "example": "jane@example.com"
                },
                "subject": {
                  "type": "string",
                  "example": "Science"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Teacher updated successfully"
          },
          "422": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["teachers"],
        "description": "Delete a teacher by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Teacher ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Teacher not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/": {
      "get": {
        "tags": ["users"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["users"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "age": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": ["users"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["users"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "age": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}